Remove deprecated commands: edit, get and toggle
Why:

 * edit, get, toggle been replaced with more uniform commands.
   Lets remove the to-be-dead code.

This change addreses the need by:

 * introduce `system history` and `entity history` as they were
   last two commands not yet moved.
 * delete edit, get and toggle plugins.

Fixes #78
